requirements.txt
amqp==2.4.2
backcall==0.1.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.0
dj-database-url==0.5.0
Django==2.2
django-filter==2.2.0
django-redis==4.10.0
djangorestframework==3.10.3
drf-yasg==1.16.1
gunicorn==19.9.0
idna==2.8
inflection==0.3.1
ipython==7.8.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.15.1
Jinja2==2.10.1
kombu==4.5.0
Markdown==3.1.1
MarkupSafe==1.1.1
packaging==19.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
prompt-toolkit==2.0.9
psycopg2==2.8.2
ptyprocess==0.6.0
Pygments==2.4.2
pyparsing==2.4.2
python-decouple==3.1
pytz==2019.1
redis==3.2.1
requests==2.21.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.1.2
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.24.2
urlwait==0.4
vine==1.3.0
wcwidth==0.1.7
whitenoise==4.1.2
